backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3
pydantic-settings==2.5.2
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
anthropic>=0.40.0
apscheduler==3.10.4
dateparser==1.2.0
twilio==9.3.5
qdrant-client>=1.7.0
openai>=1.0.0
pymupdf>=1.24.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
youtube-transcript-api>=0.6.0
yt-dlp>=2024.1.0
